refactor: migrate behaviors to unified traversal utilities
Replace `getAncestorTextBlock`, `getBlock` (from is-block), `getSpanNode`,
and `getTextBlockNode` calls in 12 behavior files with the new unified
`getTextBlock`, `getBlock`, and `getSpan` from `traversal/`.

The new utilities default to the selection focus path, eliminating
boilerplate like `getAncestorTextBlock(snapshot.context,
snapshot.context.selection.focus.path)` in favor of
`getTextBlock(snapshot)`.

The `getBlock ?? getBlock(parentPath)` fallback pattern in
`behavior.abstract.split.ts` and `behavior.abstract.keyboard.ts` is
replaced by `getBlock(snapshot)` which walks up automatically.
