+++ packages/editor/src/behaviors/behavior.abstract.keyboard.ts
@@ -1,12 +1,11 @@
 import {createKeyboardShortcut} from '@portabletext/keyboard-shortcuts'
 import {defaultKeyboardShortcuts} from '../editor/default-keyboard-shortcuts'
-import {getBlock} from '../node-traversal/is-block'
 import {getFocusInlineObject} from '../selectors/selector.get-focus-inline-object'
 import {getPreviousBlock} from '../selectors/selector.get-previous-block'
 import {isSelectionCollapsed} from '../selectors/selector.is-selection-collapsed'
 import {isSelectionExpanded} from '../selectors/selector.is-selection-expanded'
 import {isTextBlockNode} from '../slate/node/is-text-block-node'
-import {parentPath} from '../slate/path/parent-path'
+import {getBlock} from '../traversal'
 import {getBlockEndPoint} from '../utils/util.get-block-end-point'
 import {isEmptyTextBlock} from '../utils/util.is-empty-text-block'
 import {raise} from './behavior.types.action'
@@ -135,10 +134,7 @@ export const abstractKeyboardBehaviors = [
         return false
       }
 
-      const focusPath = snapshot.context.selection.focus.path
-      const focusBlock =
-        getBlock(snapshot.context, focusPath) ??
-        getBlock(snapshot.context, parentPath(focusPath))
+      const focusBlock = getBlock(snapshot)
 
       if (!focusBlock) {
         return false
